# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Gfact API",
    "version": "1.0.0",
    "description": "API de gerenciamento de subscrições do Gfact"
  },
  "servers": [
    {
      "url": "http://localhost:4002/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Subscription": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string"
          },
          "planId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "PENDING",
              "CANCELED",
              "EXPIRED"
            ]
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "renewsAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "canceledAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "paymentMethod": {
            "type": "string",
            "nullable": true
          },
          "externalRef": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/subscriptions/{id}": {
      "get": {
        "summary": "Obter uma subscrição específica",
        "description": "Busca uma subscrição pelo ID. O usuário deve ser administrador ou dono da subscrição.",
        "tags": [
          "Subscriptions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID da subscrição",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalhes da subscrição",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Subscription"
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Não autorizado"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Subscrição não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Subscrição não encontrada"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro ao buscar subscrição",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Erro ao buscar subscrição"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Atualizar uma subscrição",
        "description": "Atualiza uma subscrição existente. Apenas administradores ou donos podem atualizar.",
        "tags": [
          "Subscriptions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID da subscrição",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "planId": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string"
                  },
                  "endDate": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "renewsAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "paymentMethod": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Subscrição atualizada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Subscription"
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Não autorizado"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Subscrição não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Subscrição não encontrada"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro ao buscar subscrição",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Erro ao buscar subscrição"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Cancelar uma subscrição",
        "description": "Cancela uma subscrição existente. Apenas administradores ou donos podem cancelar.",
        "tags": [
          "Subscriptions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID da subscrição",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Subscrição cancelada com sucesso"
          },
          "403": {
            "description": "Acesso negado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Não autorizado"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Subscrição não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Subscrição não encontrada"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro ao buscar subscrição",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Erro ao buscar subscrição"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/subscriptions": {
      "post": {
        "summary": "Criar uma nova subscrição",
        "description": "Cria uma nova subscrição para o usuário autenticado.",
        "tags": [
          "Subscriptions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "planId": {
                    "type": "string"
                  },
                  "paymentMethod": {
                    "type": "string"
                  },
                  "externalRef": {
                    "type": "string"
                  },
                  "endDate": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Subscrição criada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Subscription"
                }
              }
            }
          },
          "400": {
            "description": "Já possui uma subscrição ativa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Já possui uma subscrição"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Plano não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Não encontrado"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro ao criar subscrição",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Erro inesperado"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}